feat: Add invoice management core functionality
Introduce migrations, models, and controllers for invoice details and attachments.

- Implement product fetching via AJAX in InvoicesController for the 'Add Invoice' form.
- Create the 'Add Invoice' view (add_invoice.blade.php) with required fields and VAT calculation logic.
- Add success notifications for CRUD operations ('delete', 'status update', 'restore') in the invoices index view.
- Update 'invoices' migration to use 'bigIncrements' for the primary key.

+++ database/migrations/2025_10_09_070636_create_invoices_table.php
@@ -14,7 +14,7 @@ class CreateInvoicesTable extends Migration
     public function up()
     {
         Schema::create('invoices', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('invoice_number');
             $table->date('invoice_Date');
             $table->date('due_date');
@@ -22,9 +22,9 @@ class CreateInvoicesTable extends Migration
             $table->string('section');
             $table->string('discount');
             $table->string('rate_vat');
-            $table->decimal('value_vat',8,2);
-            $table->decimal('total',8,2);
-            $table->string('status',50);
+            $table->decimal('value_vat', 8, 2);
+            $table->decimal('total', 8, 2);
+            $table->string('status', 50);
             $table->integer('value_status');
             $table->text('note')->nullable();
             $table->string('user');
